+++ ragframework/generator/__init__.py
@@ -1,5 +1,6 @@
 """Generator implementations."""
 
 from ragframework.generator.echo_generator import EchoGenerator
+from ragframework.generator.openai import OpenAIGenerator
 
-__all__ = ["EchoGenerator"]
+__all__ = ["EchoGenerator", "OpenAIGenerator"]